% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_dag_df.R
\name{make_design_dag_df}
\alias{make_design_dag_df}
\title{Title}
\usage{
make_design_dag_df(design, dag)
}
\arguments{
\item{design}{A DeclareDesign design}

\item{dag}{a}
}
\value{

}
\description{
Title
}
\examples{

}
